Monitoring: Move `QueryObj` type to dev-console package
This type is only used by the dev-console code.
