+++ frontend/packages/dev-console/src/components/monitoring/metrics/MetricsQueryInput.tsx
@@ -14,7 +14,6 @@ import {
 } from '@console/internal/actions/observe';
 import { PROMETHEUS_BASE_PATH } from '@console/internal/components/graphs';
 import { getPrometheusURL, PrometheusEndpoint } from '@console/internal/components/graphs/helpers';
-import { QueryObj } from '@console/internal/components/monitoring/query-browser';
 import {
   Dropdown,
   removeQueryArgument,
@@ -26,6 +25,7 @@ import { RootState } from '@console/internal/redux';
 import { metricsQuery, getTopMetricsQueries } from '../queries';
 import { QueryInput } from './QueryInput';
 import './MetricsQueryInput.scss';
+import { QueryObj } from './types';
 
 const ADD_NEW_QUERY = '#ADD_NEW_QUERY#';
