+++ frontend/packages/dev-console/src/components/monitoring/metrics/types.ts
@@ -1,0 +1,10 @@
+import { PrometheusLabels } from '@console/dynamic-plugin-sdk';
+
+export type QueryObj = {
+  disabledSeries?: PrometheusLabels[];
+  isEnabled?: boolean;
+  isExpanded?: boolean;
+  query?: string;
+  series?: PrometheusLabels[];
+  text?: string;
+};
